refactor managers to own sub module
