+++ unicoremc/managers/__init__.py
@@ -1,0 +1,5 @@
+from unicoremc.managers.config import ConfigManager
+from unicoremc.managers.settings import SettingsManager
+from unicoremc.managers.database import DbManager
+
+__all__ = ['ConfigManager', 'SettingsManager', 'DbManager']
